add root category if user doesn't have one

+++ app/Services/CategoryService.php
@@ -30,12 +30,21 @@ class CategoryService
         ");
         return $categories;
     }
-
+    public function addRootCategory($user_id)
+    {
+        if (!Category::whereUserId($user_id)->whereNull('parent_id')->exists())
+            return Category::create([
+                'name' => 'My rooy category',
+                'user_id' => $user_id,
+                'level' => 0
+            ]);
+        else return null;
+    }
     public function addCategory($name,  $parent_id, $desc = null, $discount_percentage = null)
     {
         $parent = Category::find($parent_id);
         if ($parent->user_id != \Auth::id())
-            throw new UnauthorizedException('UnAuthorized');
+            throw new UnauthorizedException('UnAuthorized',403);
         if ($parent->items()->exists())
             throw new BadRequestException('This category has items');
         if ($parent->level > 3)
@@ -55,7 +64,7 @@ class CategoryService
     {
         $category = Category::find($id);
         if ($category->user_id != \Auth::id())
-            throw new UnauthorizedException('UnAuthorized');
+            throw new UnauthorizedException('UnAuthorized',403);
         return $category->update($updatedData);
     }
 
@@ -63,7 +72,9 @@ class CategoryService
     {
         $category = Category::find($id);
         if ($category->user_id != \Auth::id())
-            throw new UnauthorizedException('UnAuthorized');
+            throw new UnauthorizedException('UnAuthorized',403);
+        if ($category->parent_id == null)
+            throw new UnauthorizedException('You can\'t delete the root category');
         return $category->delete();
     }
 }
